mission: fix mission download retransmission
This fixes the case where a mission item gets lost from the autopilot to
DroneCore and we need to request the same one again.

Previously, we just requested the next one, essentially skipping the one
we missed. This then lead the autopilot to respond with a failure
because the wrong sequence was requested.